refactor(cli): Restructure cli into module with subcommands
