+++ alvoc/cli/__init__.py
@@ -1,0 +1,1 @@
+from alvoc.cli.main import cli
